+++ systems/gamestates/transition.lua
@@ -78,7 +78,7 @@ function transition.update(dt)
         st.transitionPhase = nil
         st.gameState = constants.GAME_STATE_SHOP
         sound:playSegment("intro")
-        shop.abrir(st.monedas)
+        shop.abrir(st.monedas, true)
     end
 end